Update patch set 3
Patch Set 3:

Regarding ToolPath, one could make the argument that it should be removed, since it is not currently being called from anywhere. That would not be my first preference, since I have written a generator which uses this call. In these cases, I would say that some extra work probably needs to be done outside of MSVSVersion and within the generator itself, so it would make sense to either return none or raise if the MSVS version is greater than 2015. Regardless, this is something that the GYP core maintainers probably have a clearer insight on.

Just FYI, I'm not a GYP owner, so I can't merge this patch for you. But thanks for your contribution!

Patch-set: 3
CC: Gerrit User 1183044 <1183044@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>