Update patch set 3
Patch Set 3: Code-Review+1

(3 comments)

I created a subset of this CL here:

https://chromium-review.googlesource.com/c/452865/

because it is needed to fix crbug.com/700524 (well, it's the simplest way to fix it anyway). Because this CL is more complete it would be better to land it. I found one error and two minor nits, but otherwise this looks good.

Patch-set: 3
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1