Update patch set 4
Patch Set 4:

Sorry, but this is not the way to do that. Registry key approach was discussed downstream in nodejs, Heath Stewart who is responsible for the VS installer said explicitly that the registry entry is not supported (https://github.com/nodejs/node-gyp/pull/1103#issuecomment-280146014)

In node-gyp we used Rafael's powershell script that can query VS COM server (https://github.com/nodejs/node-gyp/pull/1130). We used that to find VS and set environment variables. IMHO we should do the same thing here.

Patch-set: 4