Update patch set 6
Patch Set 6:

Have the tests been run, for VS 2015 and VS 2017?

And, do we know if this fixes the clang update problem in Chromium when running gclient sync when GYP_MSVS_VERSION=2017 (see the first changed line in https://codereview.chromium.org/2748523002)?

I'm concerned about the license agreements and copyright statements in the new files. I couldn't find gyp's policy but most Google projects require that all contributions use the standard license for the project, sign a contributor agreement, etc.

Finally, I gotta say that I'm not thrilled with this change. Maybe it's necessary, but a batch file, a powershell file, and 326 lines of C# seems like a lot of code and it's not clear to me what problem it is really solving that wouldn't be solved by assuming the default install location and then using an environment variable override as an escape hatch. I replied to Microsoft's question on the other CL about why the COM solution makes me unhappy.

Patch-set: 6
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>