Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Should we close this CL? I'm not sure if you've been convinced that we don't want this or not yet :).

I think the latest thought was that `GYP` should use this CL (with the COM s**t) over 453201 as it just works, and I'm taking responsibility for the two COM scripts.
You should remember the `GYP` (as opposed to vs_toolchain.py) is a build tool and is used by many other projects.

Patch-set: 6