Update patch set 11
Patch Set 11:

This bug is about the fact that VS 2017 can be installed in any location, or multiple locations, and no longer creates an officially supported registry key pointing to an install location. There’s also the problem that an installation of VS 2017 may or may not contain C++ tools (as was the case with VS 2015.)

The solution proposed by the VS team—creating a tool called VSWhere—was inadequate for some scenarios because it would have required either caching a binary in your source enlistment or compiling code to locate the compiler. A number of people have asked for a more straightforward solution.

I’m happy to tell you that your feedback was heard. We’re going to start shipping the VSWhere tool in a known location on any machine that has VS 2017. More details can be found in this blog post: https://blogs.msdn.microsoft.com/heaths/2017/04/21/vswhere-is-now-installed-with-visual-studio-2017/

This doesn’t fix the problem for all the versions of VS 2017 that are installed already. But at least in the future you’ll be able to locate VS installs from a scripted environment.

Patch-set: 11
CC: Gerrit User 1200474 <1200474@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>