Add new LUCI tryserver, ci configs.
We want to switch GYP to use the new LUCI system to replace
buildot for the commit queue and continuous integration.

This CL adds entries to cr-buildbucket.cfg to configure the
"swarmbucket" service which is the thing that handles tryjobs
and the CQ, and adds a new luci-scheduler.cfg to replace the
GYP waterfall master. Each service will have three builders,
called "linux", "mac", and "win".

R: nodir@chromium.org, tandrii@chromium.org
Bug: gyp:539
Change-Id: I995e49978ba8cb0eeb37b0c9513a94500ac6f812